Add MasterTable
