+++ packages/devtools/devtools/src/components/index.ts
@@ -6,6 +6,7 @@ export * from './Autocomplete';
 export * from './ErrorBoundary';
 export * from './KeySelect';
 export * from './KeyTable';
+export * from './MasterTable';
 export * from './MessageTable';
 export * from './Panel';
 export * from './SpaceTable';
