+++ packages/devtools/devtools/src/containers/SignalPanel/SignalMessages.tsx
@@ -3,19 +3,12 @@
 //
 
 import React, { useEffect, useState } from 'react';
-import { Column } from 'react-table';
 
-import { Searchbar, Selector, SelectorOption, Table } from '@dxos/kai';
 import { SignalResponse } from '@dxos/protocols/proto/dxos/devtools/host';
 import { useDevtools } from '@dxos/react-client';
-import { JsonTreeView } from '@dxos/react-components-deprecated';
 import { humanize } from '@dxos/util';
 
-type ColumnType<T extends {}> = SelectorOption & {
-  filter: (object: T) => boolean;
-  subFilter: (match?: string) => (object: T) => boolean;
-  columns: Column<T>[];
-};
+import { ColumnType, MasterTable } from '../../components';
 
 const matchJsonString =
   (match = '') =>
@@ -49,8 +42,8 @@ const types: ColumnType<SignalResponse>[] = [
       {
         Header: 'Peer',
         accessor: (response: SignalResponse) =>
-          humanize(response.swarmEvent!.peerAvailable?.peer ?? '') ||
-          humanize(response.swarmEvent!.peerLeft?.peer ?? '')
+          (response.swarmEvent!.peerAvailable && humanize(response.swarmEvent!.peerAvailable.peer)) ||
+          (response.swarmEvent!.peerLeft && humanize(response.swarmEvent!.peerLeft.peer))
       },
       {
         Header: 'Since',
@@ -107,36 +100,19 @@ const types: ColumnType<SignalResponse>[] = [
         Header: 'Recipient',
         accessor: (response: SignalResponse) => humanize(response.message!.recipient)
       },
-      { Header: 'MessageID', accessor: (response: SignalResponse) => humanize(response.message?.payload.messageId) }
+      {
+        Header: 'MessageID',
+        accessor: (response: SignalResponse) => humanize(response.message!.payload.messageId)
+      }
     ]
   }
 ];
-
-const getType = (id: string): ColumnType<SignalResponse> => types.find((type) => type.id === id)!;
 
 export const SignalMessages = () => {
   const devtoolsHost = useDevtools();
   if (!devtoolsHost) {
     return null;
   }
-
-  const [text, setText] = useState<string>('');
-  const handleSearch = (text: string) => {
-    setText(text);
-  };
-
-  const [type, setType] = useState<ColumnType<SignalResponse>>(types[0]);
-  const selectType = (id?: string) => {
-    if (id) {
-      setType(getType(id));
-      setSelected(0);
-    }
-  };
-
-  const [selected, setSelected] = useState<number>(0);
-  const selectRow = (index: number) => {
-    setSelected(index);
-  };
 
   const [signalResponses, setSignalResponses] = useState<SignalResponse[]>([]);
 
@@ -153,33 +129,5 @@ export const SignalMessages = () => {
     };
   }, []);
 
-  const getFilteredData = () => signalResponses.filter(type.filter).filter(type.subFilter(text));
-
-  return (
-    <div className='flex flex-col flex-1'>
-      <div className='flex p-3 border-b border-slate-200 border-solid'>
-        <div className='flex'>
-          <div className='mr-2'>
-            <Selector options={types} value={type.id} onSelect={selectType} />
-          </div>
-          <div>
-            <Searchbar onSearch={handleSearch} />
-          </div>
-        </div>
-      </div>
-      <div className='flex flex-row'>
-        <div className='flex w-1/2 h-full'>
-          <Table
-            columns={type.columns as any}
-            data={getFilteredData() as any}
-            selected={selected}
-            onSelect={selectRow}
-          />
-        </div>
-        <div className='flex w-1/2'>
-          <JsonTreeView data={getFilteredData().at(selected)} />
-        </div>
-      </div>
-    </div>
-  );
+  return <MasterTable types={types} data={signalResponses} />;
 };
